+++ backend/src/main/java/top/phakeandy/youchat/auth/AuthenticationService.java
@@ -2,6 +2,8 @@ package top.phakeandy.youchat.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import top.phakeandy.youchat.auth.request.LoginRequest;
+import top.phakeandy.youchat.auth.response.LoginResponse;
 
 public interface AuthenticationService {
