+++ backend/src/main/java/top/phakeandy/youchat/auth/exception/AuthenticationException.java
@@ -1,26 +1,26 @@
-package top.phakeandy.youchat.auth;
+package top.phakeandy.youchat.auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
 import org.springframework.web.ErrorResponseException;
 
-public class InvalidCredentialsException extends ErrorResponseException {
+public class AuthenticationException extends ErrorResponseException {
 
   private static final long serialVersionUID = 1L;
 
-  public InvalidCredentialsException(String message) {
+  public AuthenticationException(String message) {
     super(
         HttpStatus.UNAUTHORIZED,
         ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, message),
         null);
-    getBody().setTitle("凭据无效");
+    getBody().setTitle("认证失败");
   }
 
-  public InvalidCredentialsException(String message, Throwable cause) {
+  public AuthenticationException(String message, Throwable cause) {
     super(
         HttpStatus.UNAUTHORIZED,
         ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, message),
         cause);
-    getBody().setTitle("凭据无效");
+    getBody().setTitle("认证失败");
   }
 }
