+++ backend/src/main/java/top/phakeandy/youchat/auth/exception/InvalidCredentialsException.java
@@ -1,26 +1,26 @@
-package top.phakeandy.youchat.auth;
+package top.phakeandy.youchat.auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
 import org.springframework.web.ErrorResponseException;
 
-public class AuthenticationException extends ErrorResponseException {
+public class InvalidCredentialsException extends ErrorResponseException {
 
   private static final long serialVersionUID = 1L;
 
-  public AuthenticationException(String message) {
+  public InvalidCredentialsException(String message) {
     super(
         HttpStatus.UNAUTHORIZED,
         ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, message),
         null);
-    getBody().setTitle("认证失败");
+    getBody().setTitle("凭据无效");
   }
 
-  public AuthenticationException(String message, Throwable cause) {
+  public InvalidCredentialsException(String message, Throwable cause) {
     super(
         HttpStatus.UNAUTHORIZED,
         ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, message),
         cause);
-    getBody().setTitle("认证失败");
+    getBody().setTitle("凭据无效");
   }
 }
