+++ backend/src/main/java/top/phakeandy/youchat/user/responcse/UserResponse.java
@@ -1,4 +1,4 @@
-package top.phakeandy.youchat.user;
+package top.phakeandy.youchat.user.responcse;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Collection;
