+++ backend/src/test/java/top/phakeandy/youchat/auth/AuthenticationControllerIntegrationTest.java
@@ -16,6 +16,7 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
+import top.phakeandy.youchat.auth.request.LoginRequest;
 import top.phakeandy.youchat.config.TestBase;
 
 @SpringBootTest
